+++ pyinterpolate/processing/checks.py
@@ -5,6 +5,9 @@ Authors
 -------
 1. Szymon Moliński | @SimonMolinsky
 """
+import warnings
+from typing import Collection
+from pyinterpolate.processing.utils.exceptions import SetDifferenceWarning
 
 
 def check_limits(value: float, lower_limit=0, upper_limit=1, exclusive_lower=True, exclusive_upper=True):
@@ -57,3 +60,33 @@ def check_limits(value: float, lower_limit=0, upper_limit=1, exclusive_lower=Tru
     # >
     if value > upper_limit:
         raise ValueError(msg)
+
+
+def check_ids(ids_a: Collection, ids_b: Collection, set_name_a='set 1', set_name_b='set 2'):
+    """
+    Function checks if there are any values that are missing in one set of ids, and present in another.
+
+    Parameters
+    ----------
+    ids_a : Collection
+
+    ids_b : Collection
+
+    set_name_a : str, default='set 1'
+        The name of the first set (for a warning message).
+
+    set_name_b : str, default='set 2'
+        The name of the second set (for a warning message).
+
+    Warns
+    -----
+
+    """
+    set_ids_a = set(ids_a)
+    set_ids_b = set(ids_b)
+
+    diff_a_to_b = set_ids_a.difference(set_ids_b)
+    diff_b_to_a = set_ids_b.difference(set_ids_a)
+
+    if len(diff_a_to_b) > 0 or len(diff_b_to_a) > 0:
+        warnings.warn(SetDifferenceWarning(diff_a_to_b, diff_b_to_a, set_name_a, set_name_b))